Move subpages into `/pages`, layouts out of `/pages`
