+++ client/src/components/classes/ClassesHeader.tsx
@@ -11,7 +11,7 @@ import UserDataContext, {SgyPeriod} from '../../contexts/UserDataContext';
 import SgyDataContext from '../../contexts/SgyDataContext';
 
 // Utilities
-import {findClassesList} from '../../pages/Classes';
+import {findClassesList} from './ClassesLayout';
 import {bgColor} from '../../util/progressBarColor';
